+++ components/outerTemplate/src/index.js
@@ -1,2 +1,31 @@
-const test3 = 123;
-export { test3 };
+// Get components
+
+import blobHead from "./components/blobHead";
+import bodyStyle from "./components/bodyStyle";
+import font1 from "./components/font1";
+import font2 from "./components/font2";
+import footerComponent from "./components/footerComponent";
+import headComponent from "./components/headComponent";
+import headMetaTags from "./components/headMetaTags";
+import ifStyle from "./components/ifStyle";
+import mediaStyle from "./components/mediaStyle";
+import style from "./components/style";
+import style2 from "./components/style2";
+import tableHeaderBlockLogo from "./components/tableHeaderBlockLogo";
+import topPanelViewInBrows from "./components/topPanelViewInBrows";
+
+export {
+  blobHead,
+  bodyStyle,
+  font1,
+  font2,
+  footerComponent,
+  headComponent,
+  headMetaTags,
+  ifStyle,
+  mediaStyle,
+  style,
+  style2,
+  tableHeaderBlockLogo,
+  topPanelViewInBrows,
+};
