+++ src/index.js
@@ -1,10 +1,28 @@
 // Get components
 
-// inner tamplate start
+// inner tamplate
 import { contentMainTableWrap } from "../components/innerTemplate/src/index";
-// inner tamplate end
-import { test2 } from "../components/miscellaneous/src/index";
-import { test3 } from "../components/outerTemplate/src/index";
+
+// miscellaneous tamplate (empty)
+// import { miscellaneous } from "../components/miscellaneous/src/index";
+
+// outerTemplate tamplate start
+import {
+  blobHead,
+  bodyStyle,
+  font1,
+  font2,
+  footerComponent,
+  headComponent,
+  headMetaTags,
+  ifStyle,
+  mediaStyle,
+  style,
+  style2,
+  tableHeaderBlockLogo,
+  topPanelViewInBrows,
+} from "../components/outerTemplate/src/index";
+// outerTemplate tamplate end
 
 // typography start
 import {
@@ -14,8 +32,6 @@ import {
 } from "../components/typography/src/index";
 // typography end
 
-// import headComponent from "./components/headComponent.js";
-// import footerComponent from "./components/footerComponent.js";
 import UnsubLink from "./miscellaneous/get-unsub-link.js";
 import recipe from "./components/get-recipe.js";
 import button from "./components/get-button.js";
@@ -26,11 +42,8 @@ import supportContact from "./components/get-supportContact.js";
 import content from "./components/get-content.js";
 import tableForNotes from "./components/tableForNotes.js";
 import tableWithTwoRecipes from "./components/tableWithTwoRecipes.js";
-// import contentMainTableWrap from "./outerTemplate/contentMainTableWrap.js";
 
 export {
-  // headComponent,
-  // footerComponent,
   UnsubLink,
   recipe,
   button,
@@ -41,8 +54,28 @@ export {
   content,
   tableForNotes,
   tableWithTwoRecipes,
-  // contentMainTableWrap,
+  // inner tamplate start
+  contentMainTableWrap,
+  // inner tamplate end
+
+  // outerTamplate start
+  blobHead,
+  bodyStyle,
+  font1,
+  font2,
+  footerComponent,
+  headComponent,
+  headMetaTags,
+  ifStyle,
+  mediaStyle,
+  style,
+  style2,
+  tableHeaderBlockLogo,
+  topPanelViewInBrows,
+  // outerTamplate end
+  // typography start
   headings,
   paragraph,
   strong,
+  // typography end
 };
